test check err
