+++ regression-test/data/external_table_p0/hive/ddl/test_hive_ctas.out
@@ -1,0 +1,66 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !ctas_01 --
+2
+3
+
+-- !ctas_02 --
+2
+3
+
+-- !ctas_03 --
+22	value_for_pt11	value_for_pt22
+
+-- !ctas_04 --
+11	value_for_pt1	value_for_pt2
+22	value_for_pt11	value_for_pt22
+
+-- !ctas_ex01 --
+2
+3
+
+-- !ctas_ex02 --
+11	value_for_pt1	value_for_pt2
+22	value_for_pt11	\N
+22	value_for_pt11	value_for_pt22
+
+-- !ctas_03 --
+\N	another string value for col2
+\N	string value for col2
+\N	yet another string value for col2
+
+-- !ctas_04 --
+\N	11	value_for_pt1
+\N	22	value_for_pt11
+
+-- !ctas_01 --
+2
+3
+
+-- !ctas_02 --
+2
+3
+
+-- !ctas_03 --
+22	value_for_pt11	value_for_pt22
+
+-- !ctas_04 --
+11	value_for_pt1	value_for_pt2
+22	value_for_pt11	value_for_pt22
+
+-- !ctas_ex01 --
+2
+3
+
+-- !ctas_ex02 --
+11	value_for_pt1	value_for_pt2
+22	value_for_pt11	\N
+22	value_for_pt11	value_for_pt22
+
+-- !ctas_03 --
+\N	another string value for col2
+\N	string value for col2
+\N	yet another string value for col2
+
+-- !ctas_04 --
+\N	11	value_for_pt1
+\N	22	value_for_pt11
