+++ regression-test/data/external_table_p0/hive/ddl/test_hive_ddl.out
@@ -1,0 +1,60 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !insert01 --
+true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
+
+-- !insert02 --
+\N	123	\N	\N	8.98
+true	123	9876543210	\N	stringval
+true	123	9876543210	123.4567	stringval
+
+-- !insert03 --
+\N	123	\N	\N	\N	\N	\N	varcharval	8.98
+false	1	9876543210	abcdefghij	2.3	6.28	0.0000	2223	stringval
+true	\N	9876543210	abcdefghij	2.3	6.28	\N	varcharval	stringval
+true	123	9876543210	\N	\N	\N	\N	varcharval	stringval
+true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
+
+-- !insert04 --
+true	1	1000	2.3	value_for_pt1	value_for_pt2
+
+-- !insert05 --
+true	1	1000	2.3
+true	1	1000	2.3
+true	1	1000	2.3
+
+-- !insert06 --
+\N	1	1000	1.3
+false	1	1000	\N
+true	1	1000	2.3
+true	1	1000	2.3
+true	1	1000	2.3
+
+-- !insert01 --
+true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
+
+-- !insert02 --
+\N	123	\N	\N	8.98
+true	123	9876543210	\N	stringval
+true	123	9876543210	123.4567	stringval
+
+-- !insert03 --
+\N	123	\N	\N	\N	\N	\N	varcharval	8.98
+false	1	9876543210	abcdefghij	2.3	6.28	0.0000	2223	stringval
+true	\N	9876543210	abcdefghij	2.3	6.28	\N	varcharval	stringval
+true	123	9876543210	\N	\N	\N	\N	varcharval	stringval
+true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
+
+-- !insert04 --
+true	1	1000	2.3	value_for_pt1	value_for_pt2
+
+-- !insert05 --
+true	1	1000	2.3
+true	1	1000	2.3
+true	1	1000	2.3
+
+-- !insert06 --
+\N	1	1000	1.3
+false	1	1000	\N
+true	1	1000	2.3
+true	1	1000	2.3
+true	1	1000	2.3
