--- regression-test/data/external_table_p0/hive/ddl/test_hive_ddl_and_ctas.out
@@ -1,162 +1,0 @@
--- This file is automatically generated. You should know what you did if you want to edit this
--- !insert01 --
-true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
-
--- !insert02 --
-\N	123	\N	\N	8.98
-true	123	9876543210	\N	stringval
-true	123	9876543210	123.4567	stringval
-
--- !insert03 --
-\N	123	\N	\N	\N	\N	\N	varcharval	8.98
-false	1	9876543210	abcdefghij	2.3	6.28	0.0000	2223	stringval
-true	\N	9876543210	abcdefghij	2.3	6.28	\N	varcharval	stringval
-true	123	9876543210	\N	\N	\N	\N	varcharval	stringval
-true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
-
--- !insert04 --
-true	1	1000	2.3	value_for_pt1	value_for_pt2
-
--- !insert05 --
-true	1	1000	2.3
-true	1	1000	2.3
-true	1	1000	2.3
-
--- !insert06 --
-\N	1	1000	1.3
-false	1	1000	\N
-true	1	1000	2.3
-true	1	1000	2.3
-true	1	1000	2.3
-
--- !ctas_01 --
-2
-3
-
--- !ctas_02 --
-22	value_for_pt11	value_for_pt22
-
--- !ctas_03 --
-1	string value for col2
-2	another string value for col2
-3	yet another string value for col2
-
--- !ctas_04 --
-11	value_for_pt1	value_for_pt2
-22	value_for_pt11	value_for_pt22
-
--- !ctas_05 --
-1	string value for col2
-2	another string value for col2
-3	yet another string value for col2
-
--- !ctas_06 --
-11	value_for_pt1	value_for_pt2
-22	value_for_pt11	value_for_pt22
-
--- !ctas_ex01 --
-2
-3
-
--- !ctas_ex02 --
-22	value_for_pt11	\N
-22	value_for_pt11	value_for_pt22
-
--- !ctas_03 --
-\N	another string value for col2
-\N	string value for col2
-\N	yet another string value for col2
-
--- !ctas_04 --
-\N	11	value_for_pt1
-\N	22	value_for_pt11
-
--- !complex_type01 --
-a	\N	\N	\N	\N	\N	\N	\N	\N	\N	["char1", "char2"]	["c", "d"]	["string1", "string2"]	[{1:"a"}, {2:"b"}]	{1234567890123456789:"a"}	{1234567890123456789:0.12345678}	{"key":["char1", "char2"]}	{"id": 1, "gender": 1, "name": "John Doe"}	{"scale": 123.4567, "metric": ["metric1", "metric2"]}	\N
-a	b	c	d	e	1.1	12345	0.12345678	string	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N
-a	b	c	d	e	1.1	12345	0.12345678	string	[0.001, 0.002]	["char1", "char2"]	["c", "d"]	["string1", "string2"]	[{1:"a"}, {2:"b"}]{1234567890123456789:"a"}	{1234567890123456789:0.12345678}	{"key":["char1", "char2"]}	{"id": 1, "gender": 1, "name": "John Doe"}	{"scale": 123.4567, "metric": ["metric1", "metric2"]}	{"codes": [123, 456], "props": {"key1":["char1", "char2"]}}
-
--- !complex_type02 --
-a	b	c	d	e	1.1	12345	0.12345678	string	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N
-a	b	c	d	e	1.1	12345	0.12345678	string	[0.001, 0.002]	["char1", "char2"]	["c", "d"]	["string1", "string2"]	[{1:"a"}, {2:"b"}]{1234567890123456789:"a"}	{1234567890123456789:0.12345678}	{"key":["char1", "char2"]}	{"id": 1, "gender": 1, "name": "John Doe"}	{"scale": 123.4567, "metric": ["metric1", "metric2"]}	{"codes": [123, 456], "props": {"key1":["char1", "char2"]}}
-
--- !insert01 --
-true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
-
--- !insert02 --
-\N	123	\N	\N	8.98
-true	123	9876543210	\N	stringval
-true	123	9876543210	123.4567	stringval
-
--- !insert03 --
-\N	123	\N	\N	\N	\N	\N	varcharval	8.98
-false	1	9876543210	abcdefghij	2.3	6.28	0.0000	2223	stringval
-true	\N	9876543210	abcdefghij	2.3	6.28	\N	varcharval	stringval
-true	123	9876543210	\N	\N	\N	\N	varcharval	stringval
-true	123	9876543210	abcdefghij	3.14	6.28	123.4567	varcharval	stringval
-
--- !insert04 --
-true	1	1000	2.3	value_for_pt1	value_for_pt2
-
--- !insert05 --
-true	1	1000	2.3
-true	1	1000	2.3
-true	1	1000	2.3
-
--- !insert06 --
-\N	1	1000	1.3
-false	1	1000	\N
-true	1	1000	2.3
-true	1	1000	2.3
-true	1	1000	2.3
-
--- !ctas_01 --
-2
-3
-
--- !ctas_02 --
-22	value_for_pt11	value_for_pt22
-
--- !ctas_03 --
-1	string value for col2
-2	another string value for col2
-3	yet another string value for col2
-
--- !ctas_04 --
-11	value_for_pt1	value_for_pt2
-22	value_for_pt11	value_for_pt22
-
--- !ctas_05 --
-1	string value for col2
-2	another string value for col2
-3	yet another string value for col2
-
--- !ctas_06 --
-11	value_for_pt1	value_for_pt2
-22	value_for_pt11	value_for_pt22
-
--- !ctas_ex01 --
-2
-3
-
--- !ctas_ex02 --
-22	value_for_pt11	\N
-22	value_for_pt11	value_for_pt22
-
--- !ctas_03 --
-\N	another string value for col2
-\N	string value for col2
-\N	yet another string value for col2
-
--- !ctas_04 --
-\N	11	value_for_pt1
-\N	22	value_for_pt11
-
--- !complex_type01 --
-a	\N	\N	\N	\N	\N	\N	\N	\N	\N	["char1", "char2"]	["c", "d"]	["string1", "string2"]	[{1:"a"}, {2:"b"}]	{1234567890123456789:"a"}	{1234567890123456789:0.12345678}	{"key":["char1", "char2"]}	{"id": 1, "gender": 1, "name": "John Doe"}	{"scale": 123.4567, "metric": ["metric1", "metric2"]}	\N
-a	b	c	d	e	1.1	12345	0.12345678	string	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N
-a	b	c	d	e	1.1	12345	0.12345678	string	[0.001, 0.002]	["char1", "char2"]	["c", "d"]	["string1", "string2"]	[{1:"a"}, {2:"b"}]{1234567890123456789:"a"}	{1234567890123456789:0.12345678}	{"key":["char1", "char2"]}	{"id": 1, "gender": 1, "name": "John Doe"}	{"scale": 123.4567, "metric": ["metric1", "metric2"]}	{"codes": [123, 456], "props": {"key1":["char1", "char2"]}}
-
--- !complex_type02 --
-a	b	c	d	e	1.1	12345	0.12345678	string	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N	\N
-a	b	c	d	e	1.1	12345	0.12345678	string	[0.001, 0.002]	["char1", "char2"]	["c", "d"]	["string1", "string2"]	[{1:"a"}, {2:"b"}]{1234567890123456789:"a"}	{1234567890123456789:0.12345678}	{"key":["char1", "char2"]}	{"id": 1, "gender": 1, "name": "John Doe"}	{"scale": 123.4567, "metric": ["metric1", "metric2"]}	{"codes": [123, 456], "props": {"key1":["char1", "char2"]}}
